Detecting when quests are complete. Moving on to other quests after finishing one.
