+++ eosclubhouse/libquest.py
@@ -94,6 +94,7 @@ class Quest(GObject.GObject):
         self._characters = {}
         self._main_character_id = main_character_id
         self._cancellable = None
+        self._complete = False
         self.load_conf()
 
     def start(self):
@@ -187,7 +188,10 @@ class QuestSet(GObject.GObject):
 
     def get_next_quest(self):
         quests = self.get_quests()
-        return quests[0] if quests else None
+        for quest in quests:
+            if not quest.get_conf('complete'):
+                return quest
+        return None
 
     def get_position(self):
         return self._position
